requirements.txt
blinker==1.8.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.1.1
idna==3.11
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.32.4
sqlalchemy==2.0.46
typing-extensions==4.13.2
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
